# src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Travel Buddy API",
    "version": "1.0.0",
    "description": "API documentation for Travel Buddy backend, including auth, preferences, destinations, orders, audit, trips, and itinerary."
  },
  "servers": [
    { "url": "http://localhost:{port}", "variables": { "port": { "default": "3000" } , "jwt": { "default": "Bearer <JWT>" } } }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
...
      "DestinationCreate": { "type": "object", "properties": { "country": { "type":"string" }, "city": { "type":"string" }, "bestTimeToVisit": { "type":"string" }, "avgCost": { "type":"number" }, "activities": { "type":"array","items":{"type":"string"} }, "safetyScore": { "type":"number" }, "metadata": { "type":"object" } }, "required": ["country","city"] },
...
      "TripCreate": { "type":"object","properties":{ "startDate": {"type":"string","format":"date"}, "endDate": {"type":"string","format":"date"}, "status":{"type":"string","enum":["planned","ongoing","completed"]}, "notes":{"type":"string"}, "companions":{"type":"array","items":{"type":"string"}}, "destinationIds":{"type":"array","items":{"type":"string"}} }, "required":["startDate","endDate"] },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/destinations": { "get": { "summary":"List destinations", "responses":{ "200":{"description":"List"} } }, "post": { "summary":"Create destination (admin)", "requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DestinationCreate"}}}}, "responses":{ "201":{"description":"Created"} } } },
...
    "/trips": { "post": { "summary":"Create trip", "requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TripCreate"}}}}, "responses":{ "201":{"description":"Created"} } } },
...
  }
}